feat(tachi-claim): 實作 ClaimService、ClaimHandler 與路由
- ClaimService.Claim()：跨頻道加總 spendable_balance → tachi_balances
- ClaimService.GetTachiBalance()：查詢 $TACHI 餘額
- POST /users/me/points/claim、GET /users/me/tachi/balance
- 補 service/handler 測試、兩個 testutil 加入 tachi_balances

refs #103

+++ backend/internal/models/tachi_balance.go
@@ -11,6 +11,6 @@ import (
 type TachiBalance struct {
 	ID        uuid.UUID `gorm:"type:uuid;primaryKey;default:gen_random_uuid()" json:"id"`
 	UserID    uuid.UUID `gorm:"type:uuid;not null;uniqueIndex"                 json:"user_id"`
-	Balance   float64   `gorm:"type:numeric(20,6);not null;default:0"          json:"balance"`
+	Balance   int64     `gorm:"type:numeric(20,6);not null;default:0"          json:"balance"`
 	UpdatedAt time.Time `                                                      json:"updated_at"`
 }
